# swagger/adrianoorsi-boaSaudeAPI-1.0.0-resolved.yaml
...
swagger: "2.0"
info:
  description: API do módulo MIC-GISA - Boa Saúde Assistência Médica
  version: 1.0.0
  title: API GISA - Boa Saúde
  contact:
    email: adriano.orsi1@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /adrianoorsi/boaSaudeAPI/1.0.0
...
schemes:
- https
paths:
...
  /dadosPessoais/alterar/{id}:
    post:
      tags:
      - prestadores
      - associados
      summary: alteração de dados pessoais
      description: alteração de dados pessoais de um associado
      operationId: alterarDadosPessoais
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: id do associado a alterar
        required: true
        type: integer
      - in: body
        name: dadosPessoais
        description: dados pessoais a alterar
        required: false
        schema:
          $ref: '#/definitions/DadosPessoais'
      responses:
        "201":
          description: item incluído
        "400":
          description: entrada ou objeto inválido
        "409":
          description: os dados já existem
...
  /historicoSaude/incluir:
    put:
      tags:
      - prestadores
      - associados
      summary: inclusão de Histórico de Saúde
      description: permite inclusão de item no Histórico de Saúde de um associado
      operationId: incluirHistoricoSaude
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: historicoSaude
        description: histórico de saúde a incluir
        required: false
        schema:
          $ref: '#/definitions/HistoricoSaude'
      responses:
        "201":
          description: item incluído
        "400":
          description: entrada ou objeto inválido
        "409":
          description: o histórico já existe
...
definitions:
  DadosPessoais:
    type: object
    required:
    - endereco
    - estadoCivil
    - id
    - nome
    properties:
      id:
        type: integer
        format: int32
        example: 1234
      nome:
        type: string
        example: Fulano de Tal
      endereco:
        type: string
        example: Rua Principal, 123 - Centro - CEP 12345-678 - DF - Brasil
      estadoCivil:
        type: string
        example: Solteiro
  HistoricoSaude:
    required:
    - id
    properties:
      id:
        type: integer
        format: int32
        example: 1234
      tipo:
        type: string
        example: Consulta
      dataOcorrencia:
        type: string
        format: date
        example: 26/04/2022
      dataTermino:
        type: string
        format: date
        example: 26/04/2022
      situacao:
        type: string
        example: Agendada
      CID:
        $ref: '#/definitions/CID'
  CID:
    required:
    - id
    properties:
      id:
        type: integer
        format: int32
        example: 1234
      codigo:
        type: string
        example: 6A40
      descricao:
        type: string
        example: Catatonia associada a outro transtorno mental